Update pom group_ids to match apache naming.
