+++ traffic_control/clients/java/trafficops/pom.xml
@@ -21,14 +21,14 @@
 	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<modelVersion>4.0.0</modelVersion>
 	<parent>
-		<groupId>com.comcast.cdn.traffic_control</groupId>
+		<groupId>org.apache.trafficcontrol</groupId>
 		<artifactId>traffic-control-java-client</artifactId>
 		<version>0.0.1-SNAPSHOT</version>
 	</parent>
 	<artifactId>traffic-control-java-client-trafficops</artifactId>
 	<dependencies>
 		<dependency>
-			<groupId>com.comcast.cdn.traffic_control</groupId>
+			<groupId>org.apache.trafficcontrol</groupId>
 			<artifactId>traffic-control-java-client-common</artifactId>
 			<version>${project.version}</version>
 		</dependency>
