+++ traffic_control/clients/java/trafficops/src/main/java/org/apache/trafficcontrol/client/trafficops/models/Response.java
@@ -13,7 +13,7 @@
  * limitations under the License.
  */
 
-package org.apache.trafficcontrol.models;
+package org.apache.trafficcontrol.client.trafficops.models;
 
 import java.util.List;
 import java.util.Map;
